chore: move some helper functions so that they can be shared with Sofie Core
