+++ shared/packages/api/src/index.ts
@@ -7,6 +7,7 @@ export * from './filePath'
 export * from './htmlRenderer'
 export * from './expectationApi'
 export * from './inputApi'
+export * from './inputHelpers'
 export * from './HelpfulEventEmitter'
 export * from './ids'
 export * from './dataStorage'
